move all common types to "types.d.ts"
