+++ src/frontend/constants.ts
@@ -1,15 +1,4 @@
-declare global {
-	interface Window {
-		screenControl: any;
-		onmousewheel: any;
-	}
-}
-
-interface sscOptionsDefault {
-	container: HTMLElement;
-	rootMargin: string;
-	threshold: number[];
-}
+import { SscOptionsDefault } from '../types';
 
 export const WAITFOR: number = 250;
 
@@ -22,7 +11,7 @@ export const WAITFOR: number = 250;
  * threshold: An array of numbers between 0 and 1 that defines the percentage of the target element that should be visible before it is considered visible.
  * container: The element that is used as the viewport for checking visibility of the target.
  */
-export const sscOptionsDefault: sscOptionsDefault = {
+export const sscOptionsDefault: SscOptionsDefault = {
 	rootMargin: '0px', // the IntersectionObserver root margin
 	threshold: [0],
 	container: document.body,
